+++ R/ssvs_prior.bvecmodel.R
@@ -28,7 +28,7 @@
 #' data("e6")
 #' 
 #' # Generate model input
-#' object <- create_bvecmodel(e6, r = 1)
+#' object <- create_bvecmodel(e6, p = 2, r = 1)
 #' 
 #' # Obtain SSVS prior
 #' prior <- ssvs_prior(object, semiautomatic = c(.1, 10))